+++ CHANGELOG.md
@@ -39,3 +39,16 @@
 ## v0.2.2:
 
 * updated default recipe to actually mimic LWRP reources.
+
+## v0.2.3:
+
+* added ssl tests
+* added ssl docs
+* added ssl name attribute
+* merged mwhooker to fix test resource name.
+* added ssl cert/key delete.
+
+## v0.2.4:
+
+* added limit_except (Thanks namtzigla)
+* fixed ssl delete cert issue with a new attribute